Refactor: extract DFTU atom-pair HR/occ kernels as free functions
Move DFTU<OperatorLCAO<TK,TR>>::cal_HR_IJR and cal_occ out of the
class template into namespace DFTU_LCAO (new header dftu_nao_ijr.h):

- cal_hr_ijr<TR>: TK was unused, only TR (double or complex<double>)
  is needed; npol is passed explicitly instead of reading ucell.
- cal_occ_ijr: fully non-template (all-double signature), removing
  three redundant instantiations; npol is passed explicitly.
- Update the single call site of each kernel and related comments.
- Fix __DEBUG assertion in the occ kernel: occ.size() equals
  m_size^2*npol^2 (4*m_size^2 for NSPIN=4), not m_size^2.

Behavior is unchanged in release builds.
