+++ source/source_lcao/module_dftu/dftu_nao_op_kernel.cpp
@@ -15,117 +15,8 @@
 // Include the free function implementations for force/stress in real space
 #include "dftu_nao_fs_r.h"
 
-// cal_HR_IJR: accumulate HR += <psi_I|beta_m> * pot_onsite(m,m') * <beta_m'|psi_{J,R}>
-template <typename TK, typename TR>
-void hamilt::DFTU<hamilt::OperatorLCAO<TK, TR>>::cal_HR_IJR(
-    const int& iat1,
-    const int& iat2,
-    const Parallel_Orbitals* pv,
-    const std::unordered_map<int, std::vector<double>>& nlm1_all,
-    const std::unordered_map<int, std::vector<double>>& nlm2_all,
-    const std::vector<TR>& pot_onsite,
-    TR* data_pointer)
-{
-    // npol is the number of polarizations,
-    // 1 for non-magnetic (one Hamiltonian matrix only has spin-up or spin-down),
-    // 2 for magnetic (one Hamiltonian matrix has both spin-up and spin-down)
-    const int npol = this->ucell->get_npol();
-    auto row_indexes = pv->get_indexes_row(iat1);
-    auto col_indexes = pv->get_indexes_col(iat2);
-    const int m_size = int(sqrt(pot_onsite.size()) / npol);
-    // step_trace = 0 for NSPIN=1,2; ={0, 1, local_col, local_col+1} for NSPIN=4
-    std::vector<int> step_trace(npol * npol, 0);
-    for (int is = 0; is < npol; is++)
-    {
-        for (int is2 = 0; is2 < npol; is2++)
-        {
-            step_trace[is * npol + is2] = pv->get_ncol_atom(iat2) * is + is2;
-        }
-    }
-    // calculate the local matrix
-    for (int iw1l = 0; iw1l < row_indexes.size(); iw1l += npol)
-    {
-        const std::vector<double>& nlm1 = nlm1_all.find(row_indexes[iw1l])->second;
-        for (int iw2l = 0; iw2l < col_indexes.size(); iw2l += npol)
-        {
-            const std::vector<double>& nlm2 = nlm2_all.find(col_indexes[iw2l])->second;
-#ifdef __DEBUG
-            assert(nlm1.size() == nlm2.size());
-#endif
-            for (int is = 0; is < npol * npol; ++is)
-            {
-                int start = is * m_size * m_size;
-                TR nlm_tmp = TR(0);
-                for (int m1 = 0; m1 < m_size; m1++)
-                {
-                    for (int m2 = 0; m2 < m_size; m2++)
-                    {
-                        nlm_tmp += nlm1[m1] * nlm2[m2] * pot_onsite[m1 * m_size + m2 + start];
-                    }
-                }
-                data_pointer[step_trace[is]] += nlm_tmp;
-            }
-            data_pointer += npol;
-        }
-        data_pointer += (npol - 1) * col_indexes.size();
-    }
-}
-
-// cal_occ: compute occ_mm' = sum_R DMR*<phi_0|alpha^I_m'><alpha^I_m'|phi_R>
-template <typename TK, typename TR>
-void hamilt::DFTU<hamilt::OperatorLCAO<TK, TR>>::cal_occ(const int& iat1,
-                                                         const int& iat2,
-                                                         const Parallel_Orbitals* pv,
-                                                         const std::unordered_map<int, std::vector<double>>& nlm1_all,
-                                                         const std::unordered_map<int, std::vector<double>>& nlm2_all,
-                                                         const double* dm_pointer,
-                                                         std::vector<double>& occ)
-{
-    const int npol = this->ucell->get_npol();
-    auto row_indexes = pv->get_indexes_row(iat1);
-    auto col_indexes = pv->get_indexes_col(iat2);
-    const int m_size = int(sqrt(occ.size()) / npol);
-    const int m_size2 = m_size * m_size;
-#ifdef __DEBUG
-    assert(m_size * m_size == occ.size());
-#endif
-    // step_trace = 0 for NSPIN=1,2; ={0, 1, local_col, local_col+1} for NSPIN=4
-    std::vector<int> step_trace(npol * npol, 0);
-    for (int is = 0; is < npol; is++)
-    {
-        for (int is2 = 0; is2 < npol; is2++)
-        {
-            step_trace[is * npol + is2] = pv->get_ncol_atom(iat2) * is + is2;
-        }
-    }
-    for (int iw1l = 0; iw1l < row_indexes.size(); iw1l += npol)
-    {
-        const std::vector<double>& nlm1 = nlm1_all.find(row_indexes[iw1l])->second;
-        for (int iw2l = 0; iw2l < col_indexes.size(); iw2l += npol)
-        {
-            const std::vector<double>& nlm2 = nlm2_all.find(col_indexes[iw2l])->second;
-#ifdef __DEBUG
-            assert(nlm1.size() == nlm2.size());
-#endif
-            for (int is1 = 0; is1 < npol; ++is1)
-            {
-                for (int is2 = 0; is2 < npol; ++is2)
-                {
-                    for (int m1 = 0; m1 < m_size; ++m1)
-                    {
-                        for (int m2 = 0; m2 < m_size; ++m2)
-                        {
-                            occ[m1 * m_size + m2 + (is1 * npol + is2) * m_size2]
-                                += nlm1[m1] * nlm2[m2] * dm_pointer[step_trace[is1 * npol + is2]];
-                        }
-                    }
-                }
-            }
-            dm_pointer += npol;
-        }
-        dm_pointer += (npol - 1) * col_indexes.size();
-    }
-}
+// The real-space atom-pair kernels cal_hr_ijr<TR>() and cal_occ_ijr() live
+// in dftu_nao_ijr.h as free functions of namespace DFTU_LCAO.
 
 // transfer_pot_onsite (generic: identity copy)
 template <typename TK, typename TR>
